pyproject.toml
[project]
name = "welp-app"
version = "0.1.0"
description = "Welp Application - Django project with Desk and Payflow modules"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref",
    "django",
    "pip",
    "sqlparse",
    "tzdata",
    "gunicorn",
    "boto3",
    "psycopg2-binary",
    "django-storages",
    "loguru",
    "pytest",
    "pytest-django",
    "python-dotenv",
    "django-vite",
    "django-htmx",
    "django-components",
    "Pillow",
    "PyYAML",
    "django-extensions",
    "ipython",
    "djangorestframework",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "django-extensions",
    "ipython",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "project", "welp_desk", "welp_payflow", "api"]
include = [
    "manage.py",
    "templates/**/*",
    "static/**/*",
    "scripts/**/*",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--tb=short"

[tool.django]
settings_module = "project.settings" 